ref(bin): move publicKeyToAddr out
